+++ Assets/Audio/Script/Editor/AudioSourcePausableEditor.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEditor;
+
+[CustomEditor(typeof(AudioSourcePausable))]
+public class AudioSourcePausableEditor : Editor
+{
+    public override void OnInspectorGUI()
+    {
+        
+    }
+}