+++ LocalPackages/net.peeweek.gameplay-ingredients/Editor/MenuItems.cs
@@ -142,5 +142,28 @@ namespace GameplayIngredients.Editor
         }
 
         #endregion
+
+        #region HELP
+
+        [MenuItem("Help/Gameplay Ingredients/Documentation")]
+        static void Help()
+        {
+            Application.OpenURL("https://peeweek.readthedocs.io/en/latest/gameplay-ingredients/");
+        }
+
+        [MenuItem("Help/Gameplay Ingredients/GitHub Repository (Issues and Releases)")]
+        static void GitHub()
+        {
+            Application.OpenURL("https://github.com/peeweek/net.peeweek.gameplay-ingredients/");
+        }
+
+        [MenuItem("Help/Gameplay Ingredients/OpenUPM page")]
+        static void OpenUPM()
+        {
+            Application.OpenURL("https://openupm.com/packages/net.peeweek.gameplay-ingredients/");
+        }
+
+
+        #endregion
     }
 }