+++ LocalPackages/net.peeweek.gameplay-ingredients/Runtime/LevelScripting/Actions/FocusUIAction.cs
@@ -11,6 +11,8 @@ namespace GameplayIngredients.Actions
         {
             if (UIObjectToFocus != null)
             {
+                // Workaround : Before selecting, we ensure that there's no selection in the EventSystem
+                Manager.Get<UIEventManager>().eventSystem.SetSelectedGameObject(null);
                 UIObjectToFocus.Select();
             }
         }