+++ LocalPackages/net.peeweek.gameplay-ingredients/Runtime/LevelScripting/Events/OnStartEvent.cs
@@ -2,6 +2,7 @@ using NaughtyAttributes;
 
 namespace GameplayIngredients.Events
 {
+    [AdvancedHierarchyIcon("Packages/net.peeweek.gameplay-ingredients/Icons/Events/ic-event-start.png")]
     public class OnStartEvent : EventBase
     {
         [ReorderableList]