+++ LocalPackages/net.peeweek.gameplay-ingredients/Runtime/LevelScripting/Events/OnTriggerEvent.cs
@@ -3,6 +3,7 @@ using UnityEngine;
 
 namespace GameplayIngredients.Events
 {
+    [AdvancedHierarchyIcon("Packages/net.peeweek.gameplay-ingredients/Icons/Events/ic-event-trigger.png")]
     public class OnTriggerEvent : EventBase
     {
         [ReorderableList]